+++ src/GitHub_Updater/Rest_Update.php
@@ -193,10 +193,11 @@ class Rest_Update extends Base {
 				$tag = $_REQUEST['committish'];
 			}
 
-			$current_branch = $this->get_local_branch();
 			$this->get_webhook_source();
-			if ( $tag !== $current_branch ) {
-				throw new \UnexpectedValueException( 'Request tag and webhook are not matching.' );
+			$current_branch = $this->get_local_branch();
+			$override       = isset( $_REQUEST['override'] );
+			if ( $tag !== $current_branch && ! $override ) {
+				throw new \UnexpectedValueException( 'Request tag and webhook are not matching. Consider using `override` query arg.' );
 			}
 
 			if ( isset( $_REQUEST['plugin'] ) ) {